Make subscription responses consistent
Before this change addresses which had no transactions would not send
anything to the channel after the subscription whereas addresses with
transactions would send the hash of the last transaction. This made it
difficult to use this library to monitor incoming transactions, for
example.

With this change you can always count on a reply. A hash of the last
transaction or an empty string if there are no past transactions.